Actualiazción de tipos de dato de modelos de API para mantenerlos en sincronía con BD.

+++ Kms Cloud Api/Models/ResponseModels/DataDistanceResponse.cs
@@ -7,7 +7,7 @@ namespace Kms.Cloud.Api.Models.ResponseModels {
     public class DataDistanceResponse {
         public DateTime Timestamp { get; set; }
 
-        public long Distance { get; set; }
+        public double Distance { get; set; }
         public long Steps { get; set; }
     }
 }
